updates after merging

+++ app/src/main/java/com/example/androidcourseshpp/data/network/entity/contacts/ContactData.kt
@@ -1,3 +1,3 @@
 package com.example.androidcourseshpp.data.network.entity.contacts
 
-data class ContactData(val userId: Int, val contactId: Int)
+data class ContactData(val userId: Long, val contactId: Long)
