+++ app/src/main/java/com/example/androidcourseshpp/data/network/service/user/UserService.kt
@@ -6,8 +6,8 @@ import com.example.androidcourseshpp.data.network.entity.user.UpdateUserData
 
 interface UserService {
 
-    suspend fun updateUserInfo(userId: Int, userData: UpdateUserData)
-    suspend fun getUser(userId: Int): GetUserResponseEntity
+    suspend fun updateUserInfo(userId: Long, userData: UpdateUserData)
+    suspend fun getUser(userId: Long): GetUserResponseEntity
 
     suspend fun getUsers() : GetUsersResponseEntity
 }
