+++ app/src/main/java/com/example/androidcourseshpp/data/userdata/UserDataProvider.kt
@@ -1,13 +1,12 @@
 package com.example.androidcourseshpp.data.userdata
 
 interface UserDataProvider {
-    fun saveUserServerId(userServerId: Int)
-    fun getUserServerId(): Int
+    fun saveUserServerId(userServerId: Long)
+    fun getUserServerId(): Long
     fun clearUserServerId()
     fun saveUserAvatarUrl(avatar: String)
     fun getUserAvatarUrl(): String
     fun clearUserAvatarUrl()
-
     fun saveUserGalleryPhotos(photoURLs: Set<String>)
     fun getUserGalleryPhotos(): Set<String>
 }
