+++ app/src/main/java/com/example/androidcourseshpp/ui/screens/auth/signin/SignInContract.kt
@@ -18,7 +18,7 @@ class SignInContract {
     sealed interface Effect : ViewEffect {
         data object NavigateToSingUpScreen : Effect
 
-        data class NavigateToUserProfileScreen(val userServerId: Int) : Effect
+        data class NavigateToUserProfileScreen(val userServerId: Long) : Effect
 
         data class ShowToast(val toastMessageResId: Int) : Effect
     }
