+++ app/src/main/java/com/example/androidcourseshpp/ui/screens/auth/signupextended/SignUpExtendedContract.kt
@@ -22,7 +22,7 @@ class SignUpExtendedContract {
         data object NavigateToPreviousScreen : Effect
         data object NavigateToChooseProfilePhotoDialog : Effect
         data class ShowToast(val toastMessageResId: Int) : Effect
-        data class NavigateToUserProfileScreen(val userServerId: Int) : Effect
+        data class NavigateToUserProfileScreen(val userServerId: Long) : Effect
     }
 
     data class UIState(
